Move temps into show
