+++ resources/views/livewire/printers/partials/temps.blade.php
